+++ crates/ruma-client-api/src/profile/set_avatar_url.rs
@@ -102,7 +102,7 @@ pub mod v3 {
 
     #[cfg(all(test, feature = "server"))]
     mod tests {
-        use ruma_common::api::IncomingRequest as _;
+        use ruma_common::api::IncomingRequestExt as _;
 
         use super::Request;
 
@@ -126,7 +126,11 @@ pub mod v3 {
                     http::Request::builder()
                         .method("PUT")
                         .uri("https://bar.org/_matrix/client/r0/profile/@foo:bar.org/avatar_url")
-                        .body(serde_json::to_vec(&serde_json::json!({ "avatar_url": "" })).unwrap())
+                        .body(
+                            serde_json::to_vec(&serde_json::json!({ "avatar_url": "" }))
+                                .unwrap()
+                                .as_slice(),
+                        )
                         .unwrap(),
                     &["@foo:bar.org"],
                 )