+++ crates/ruma-identity-service-api/src/association/check_3pid_validity.rs
@@ -10,15 +10,17 @@ pub mod v2 {
     use js_int::UInt;
     use ruma_common::{
         OwnedClientSecret, OwnedSessionId,
-        api::{auth_scheme::AccessToken, request, response},
+        api::{request, response},
         metadata,
         thirdparty::Medium,
     };
 
+    use crate::IdentityServiceToken;
+
     metadata! {
         method: GET,
         rate_limited: false,
-        authentication: AccessToken,
+        authentication: IdentityServiceToken,
         history: {
             1.0 => "/_matrix/identity/v2/3pid/getValidated3pid/",
         }